Abort loading of the page aborted request.
Changed $.get to $.ajax in loadItems function. Added xhrPool array at the bottom. jqXHR is saved as array and destroyed if another call is made during the page loading. This fixes my problem when I use infinite ajax scroll in a product filtering page which is loaded by Ajax. User can change the filter settings when the page is getting loaded so it needs to actually load the page from the last filter state aborting previous call.

This may not be the right way to deal my problem but this fixes my problem.